requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
supabase==2.9.1
stripe==11.1.1
jinja2==3.1.4
python-multipart==0.0.12
anthropic==0.39.0
openai>=1.0.0
